Update medusa_markDups.sh
With the somewhat recent changes to how much memory is auto assigned to 2G instead of 3G, the jobs for this are being submitted with 16G of memory, but java is told that it has 22G to play with, thus causing out of memory errors. 3G is likely a minimum, we may want to consider bumping this to 4G and then allocating 30G for java to play with.